requirements.txt
kiteconnect
google-generativeai
pandas
requests
streamlit